refactor building usage attribute validator into per-generation strategy
