+++ engine/runtime/action-plateau-processor/src/plateau6/mapping.rs
@@ -14,6 +14,7 @@ use crate::common::solid_intersection_test_pair_creator::SolidIntersectionTestPa
 use crate::common::udx_folder_extractor::UDXFolderExtractorFactory;
 use crate::common::unmatched_xlink_detector::UnmatchedXlinkDetectorFactory;
 
+use super::building_usage_attribute_strategy::Plateau6BuildingUsageStrategy;
 use super::unmatched_xlink_strategy::Plateau6XlinkStrategy;
 
 pub(crate) static ACTION_FACTORY_MAPPINGS: Lazy<HashMap<String, NodeKind>> = Lazy::new(|| {
@@ -22,7 +23,10 @@ pub(crate) static ACTION_FACTORY_MAPPINGS: Lazy<HashMap<String, NodeKind>> = Laz
         Box::new(DomainOfDefinitionValidatorFactory::new(&PLATEAU6)),
         Box::new(ObjectListExtractorFactory::new(&PLATEAU6)),
         Box::new(MissingAttributeDetectorFactory::new(&PLATEAU6)),
-        Box::new(BuildingUsageAttributeValidatorFactory::new(&PLATEAU6)),
+        Box::new(BuildingUsageAttributeValidatorFactory::new(
+            &PLATEAU6,
+            &Plateau6BuildingUsageStrategy,
+        )),
         Box::new(DestinationMeshCodeExtractorFactory::new(&PLATEAU6)),
         Box::new(BuildingPartConnectivityCheckerFactory::new(&PLATEAU6)),
         Box::new(SolidIntersectionTestPairCreatorFactory::new(&PLATEAU6)),
