+++ engine/runtime/action-plateau-processor/src/plateau6/mod.rs
@@ -1,3 +1,4 @@
+mod building_usage_attribute_strategy;
 mod mapping;
 mod profile;
 mod unmatched_xlink_strategy;
